Integrated new scheduling algorithm. Massively added try catch debugging across the scheduler. Initial testing complete, fixed all errors found with standard scheduling. Currently overtime is unhandled. Currently undertime is ignored.
