+++ Assembly-CSharp.csproj
@@ -51,7 +51,7 @@
      <Compile Include="Assets\System\EmpListSerializer.cs" />
      <Compile Include="Assets\System\EmployeeStorage.cs" />
      <Compile Include="Assets\System\FileManager.cs" />
-     <Compile Include="Assets\System\NewSchedAlgo.cs" />
+     <Compile Include="Assets\System\RetiredSchedAlgo.cs" />
      <Compile Include="Assets\System\SchedulingAlgorithm.cs" />
      <Compile Include="Assets\System\SubSystem\CurrentTime.cs" />
      <Compile Include="Assets\System\SubSystem\LoadingSpin.cs" />
