+++ VZScheduler.userprefs
@@ -1,14 +1,14 @@
 ﻿<Properties StartupItem="Assembly-CSharp.csproj">
   <MonoDevelop.Ide.Workspace ActiveConfiguration="Debug" PreferredExecutionTarget="Unity.Instance.Unity Editor" />
-  <MonoDevelop.Ide.Workbench ActiveDocument="Assets\System\CoreSystem.cs">
+  <MonoDevelop.Ide.Workbench ActiveDocument="Assets\System\SchedulingAlgorithm.cs">
     <Files>
-      <File FileName="Assets\System\CoreSystem.cs" Line="144" Column="1" />
+      <File FileName="Assets\System\SchedulingAlgorithm.cs" Line="35" Column="1" />
+      <File FileName="Assets\System\CoreSystem.cs" Line="1" Column="1" />
       <File FileName="Assets\WindowScripts\AvailChangePrompt.cs" Line="1" Column="1" />
       <File FileName="Assets\System\FileManager.cs" Line="1" Column="1" />
       <File FileName="Assets\WindowScripts\EmpSelectionWindow.cs" Line="1" Column="1" />
       <File FileName="Assets\WindowScripts\ScheduleWindow.cs" Line="1" Column="1" />
       <File FileName="Assets\WindowScripts\AvailabilityWindow.cs" Line="1" Column="1" />
-      <File FileName="Assets\System\SchedulingAlgorithm.cs" Line="1" Column="1" />
       <File FileName="Assets\System\SubSystem\LoadingSpin.cs" Line="1" Column="1" />
       <File FileName="Assets\WindowScripts\Window.cs" Line="1" Column="1" />
       <File FileName="Assets\WindowScripts\DailyStaffing.cs" Line="1" Column="1" />
